Get the new setup flow going, with a confirmation dialog etc.
Moving the folder while already installed still crashes.

+++ Core/FileLoaders/DiskCachingFileLoader.cpp
@@ -754,9 +754,9 @@ u64 DiskCachingFileLoaderCache::FreeDiskSpace() {
 		dir = GetSysDirectory(DIRECTORY_CACHE);
 	}
 
-	uint64_t result = 0;
+	int64_t result = 0;
 	if (free_disk_space(dir, result)) {
-		return result;
+		return (u64)result;
 	}
 
 	// We can't know for sure how much is free, so we have to assume none.
